+++ Sources/CodexBarCLI/CLIDiagnoseCommand.swift
@@ -239,6 +239,8 @@ extension CodexBarCLI {
             GroqSettingsReader.apiKey(environment: environment) != nil
         case .kilo:
             KiloSettingsReader.apiKey(environment: environment) != nil
+        case .factory:
+            FactorySettingsReader.apiKey(environment: environment) != nil
         default:
             false
         }